se corrigieron todos los endpoints, se genero un lector de badRequest y se dockerizo la aplicaion

+++ gfp/src/main/java/gestion/finanzas/personales/gfp/service/PresupuestoServiceImpl.java
@@ -2,7 +2,8 @@ package gestion.finanzas.personales.gfp.service;
 
 import gestion.finanzas.personales.gfp.dto.PresupuestoRequestDto;
 import gestion.finanzas.personales.gfp.dto.PresupuestoResponseDto;
-import gestion.finanzas.personales.gfp.exception.PresupuestoNoEncontrado;
+import gestion.finanzas.personales.gfp.exception.ExceptionCategoriaNotFound;
+import gestion.finanzas.personales.gfp.exception.ExceptionPresupuestoNotFound;
 import gestion.finanzas.personales.gfp.model.Categoria;
 import gestion.finanzas.personales.gfp.model.Usuario;
 import gestion.finanzas.personales.gfp.repository.PresupuestoRepository;
@@ -30,6 +31,7 @@ public class PresupuestoServiceImpl implements PresupuestoService {
     @Override
     public PresupuestoResponseDto save(PresupuestoRequestDto presupuestoRequestDto, Usuario usuario) {
         Categoria categoria = categoriaService.findById(presupuestoRequestDto.idCategoria());
+        if(Objects.isNull(categoria)) throw new ExceptionCategoriaNotFound();
         Presupuesto presupuesto = Presupuesto.builder()
                 .mes(presupuestoRequestDto.mes())
                 .categoria(categoria)
@@ -41,17 +43,21 @@ public class PresupuestoServiceImpl implements PresupuestoService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id,Usuario usuario) {
+        Presupuesto presupuesto = presupuestoRepository.findById(id).orElseThrow(ExceptionPresupuestoNotFound::new);
+        if (presupuesto.getUsuario().getId().equals(usuario.getId())) {
+            throw new ExceptionPresupuestoNotFound();
+        }
         presupuestoRepository.deleteById(id);
     }
 
     @Override
     public Presupuesto findById(Long id, Usuario usuario) {
 
-        Presupuesto presupuesto =  presupuestoRepository.findById(id).orElseThrow(PresupuestoNoEncontrado::new);
+        Presupuesto presupuesto =  presupuestoRepository.findById(id).orElseThrow(ExceptionPresupuestoNotFound::new);
         boolean perteneceAlUsuacio = Objects.equals(presupuesto.getUsuario().getId(),usuario.getId());
         if (!perteneceAlUsuacio){
-            throw new PresupuestoNoEncontrado();
+            throw new ExceptionPresupuestoNotFound();
         }
         return presupuesto;
     }
